updated test based on changes
